+++ Changelog.md
@@ -1,5 +1,6 @@
 --------------------------Changelog for the next version of SecurityCraft--------------------------
 
+- New: Reinforced block variants for all blocks in the new Sulfur and Cinnabar block sets
 - New: Secure Trading Station
 - New: The Keycard Reader, Keypad, and Retinal Scanner can now be placed facing up or down (Thanks CYB3RCA4T!)
 - New: The "/sc owner" subcommand now has an optional "resetSettings" mode to force affected blocks to reset sensible data like passcodes
@@ -13,6 +14,7 @@
 - New: Inventory Scanners will now react to items in the equipment slots of living entities
 - New: Inventory Scanners will now react to items equipped in the player's Curio inventory, if the mod Curios is installed
 - New: The Username Logger now supports the Smart Module, allowing it to update existing entries instead of creating new ones for the same player 
+- New: Dutch translation
 - Change: Several entries within the SecurityCraft Manual have been updated to remove outdated descriptions
 - Change: Ownable blocks that are targeted by the "/sc owner" subcommand will now keep all of their non-owner-related data by default
 - Change: The visible chunks in Frame feeds are no longer expanded by Distant Horizons, to increase stability and performance